+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 
 * [#2826](https://github.com/clojure-emacs/cider/pull/2826): Add support for symbols with quotes and resolving of ns-aliased keywords in `cider-symbol-at-point`.
 * [#2617](https://github.com/clojure-emacs/cider/pull/2617): Add menu bar entry for `Insert last sexp in REPL`.
+* Removed support for the Nashorn ClojureScript REPL. (it was removed upstream in ClojureScript)
 
 ### Bugs fixed